[#24]칵테일 엔티티 생성
칵테일 조주법, 칵테일 재료, 조주법별 재료 도메인 구분 및 생성
