--- src/main/java/com/chainsmoker/marronnier/cocktail/command/application/controller/CocktailController.java
@@ -1,4 +1,0 @@
-package com.chainsmoker.marronnier.cocktail.command.application.controller;
-
-public class CocktailController {
-}
