--- src/main/java/com/chainsmoker/marronnier/cocktail/command/application/dto/CocktailDTO.java
@@ -1,4 +1,0 @@
-package com.chainsmoker.marronnier.cocktail.command.application.dto;
-
-public class CocktailDTO {
-}
