--- src/main/java/com/chainsmoker/marronnier/cocktail/command/application/service/CocktailService.java
@@ -1,4 +1,0 @@
-package com.chainsmoker.marronnier.cocktail.command.application.service;
-
-public class CocktailService {
-}
