--- src/main/java/com/chainsmoker/marronnier/cocktail/command/domain/aggregate/entity/EnumType/CategoryNameEnum.java
@@ -1,8 +1,0 @@
-package com.chainsmoker.marronnier.cocktail.command.domain.aggregate.entity.EnumType;
-
-public enum CategoryNameEnum {
-
-    DRINK,
-    FRUIT,
-    ETC
-}
