--- src/main/java/com/chainsmoker/marronnier/cocktail/command/domain/aggregate/entity/EnumType/DifficultyEnum.java
@@ -1,7 +1,0 @@
-package com.chainsmoker.marronnier.cocktail.command.domain.aggregate.entity.EnumType;
-
-public enum DifficultyEnum {
-    HARD,
-    NORMAL,
-    EASY
-}
