--- src/main/java/com/chainsmoker/marronnier/cocktail/command/domain/aggregate/vo/CategoryVO.java
@@ -1,11 +1,0 @@
-package com.chainsmoker.marronnier.cocktail.command.domain.aggregate.vo;
-
-import com.chainsmoker.marronnier.cocktail.command.domain.aggregate.entity.EnumType.CategoryNameEnum;
-
-
-public class CategoryVO {
-    private Long code;
-
-    private CategoryNameEnum name;
-
-}
