--- src/main/java/com/chainsmoker/marronnier/cocktail/command/domain/repository/CocktailRepository.java
@@ -1,8 +1,0 @@
-package com.chainsmoker.marronnier.cocktail.command.domain.repository;
-
-import com.chainsmoker.marronnier.cocktail.command.domain.aggregate.entity.Cocktail;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CocktailRepository extends JpaRepository<Cocktail,Long> {
-}
-
