--- src/main/java/com/chainsmoker/marronnier/cocktail/command/domain/service/CocktailService.java
@@ -1,4 +1,0 @@
-package com.chainsmoker.marronnier.cocktail.command.domain.service;
-
-public class CocktailService {
-}
