--- src/main/java/com/chainsmoker/marronnier/cocktail/command/infra/repository/CocktailInfraRepo.java
@@ -1,4 +1,0 @@
-package com.chainsmoker.marronnier.cocktail.command.infra.repository;
-
-public class CocktailInfraRepo {
-}
