--- src/main/java/com/chainsmoker/marronnier/cocktail/command/infra/service/CocktailInfraService.java
@@ -1,4 +1,0 @@
-package com.chainsmoker.marronnier.cocktail.command.infra.service;
-
-public class CocktailInfraService {
-}
