--- src/main/java/com/chainsmoker/marronnier/cocktail/query/application/controller/CocktailQController.java
@@ -1,4 +1,0 @@
-package com.chainsmoker.marronnier.cocktail.query.application.controller;
-
-public class CocktailQController {
-}
