--- src/main/java/com/chainsmoker/marronnier/cocktail/query/application/dto/CocktailQDTO.java
@@ -1,4 +1,0 @@
-package com.chainsmoker.marronnier.cocktail.query.application.dto;
-
-public class CocktailQDTO {
-}
