--- src/main/java/com/chainsmoker/marronnier/cocktail/query/application/service/CocktailQService.java
@@ -1,4 +1,0 @@
-package com.chainsmoker.marronnier.cocktail.query.application.service;
-
-public class CocktailQService {
-}
