--- src/main/java/com/chainsmoker/marronnier/cocktail/query/domain/repository/CocktailQRepo.java
@@ -1,4 +1,0 @@
-package com.chainsmoker.marronnier.cocktail.query.domain.repository;
-
-public class CocktailQRepo {
-}
