--- src/main/java/com/chainsmoker/marronnier/cocktail/query/domain/service/CocktailQService.java
@@ -1,4 +1,0 @@
-package com.chainsmoker.marronnier.cocktail.query.domain.service;
-
-public class CocktailQService {
-}
