--- src/main/java/com/chainsmoker/marronnier/cocktail/query/infra/repository/CocktailQInfraRepo.java
@@ -1,4 +1,0 @@
-package com.chainsmoker.marronnier.cocktail.query.infra.repository;
-
-public class CocktailQInfraRepo {
-}
