--- src/main/java/com/chainsmoker/marronnier/cocktail/query/infra/service/CocktailQInfraService.java
@@ -1,4 +1,0 @@
-package com.chainsmoker.marronnier.cocktail.query.infra.service;
-
-public class CocktailQInfraService {
-}
