+++ src/main/java/com/chainsmoker/marronnier/cocktailrecipe/command/domain/aggregate/EnumType/DifficultyEnum.java
@@ -1,0 +1,7 @@
+package com.chainsmoker.marronnier.cocktailrecipe.command.domain.aggregate.EnumType;
+
+public enum DifficultyEnum {
+    HARD,
+    NORMAL,
+    EASY
+}
