+++ src/main/java/com/chainsmoker/marronnier/element/command/domain/aggregate/EnumType/CategoryNameEnum.java
@@ -1,0 +1,8 @@
+package com.chainsmoker.marronnier.element.command.domain.aggregate.EnumType;
+
+public enum CategoryNameEnum {
+
+    DRINK,
+    FRUIT,
+    ETC
+}
