+++ src/main/java/com/chainsmoker/marronnier/element/command/domain/aggregate/vo/CategoryVO.java
@@ -1,0 +1,10 @@
+package com.chainsmoker.marronnier.element.command.domain.aggregate.vo;
+
+import com.chainsmoker.marronnier.element.command.domain.aggregate.EnumType.CategoryNameEnum;
+
+public class CategoryVO {
+    private Long code;
+
+    private CategoryNameEnum name;
+
+}
