+++ src/main/java/com/chainsmoker/marronnier/elementbyrecipe/command/application/controller/CocktailController.java
@@ -1,0 +1,4 @@
+package com.chainsmoker.marronnier.elementbyrecipe.command.application.controller;
+
+public class CocktailController {
+}
