+++ src/main/java/com/chainsmoker/marronnier/elementbyrecipe/command/application/dto/CocktailDTO.java
@@ -1,0 +1,4 @@
+package com.chainsmoker.marronnier.elementbyrecipe.command.application.dto;
+
+public class CocktailDTO {
+}
