+++ src/main/java/com/chainsmoker/marronnier/elementbyrecipe/command/application/service/CocktailService.java
@@ -1,0 +1,4 @@
+package com.chainsmoker.marronnier.elementbyrecipe.command.application.service;
+
+public class CocktailService {
+}
