+++ src/main/java/com/chainsmoker/marronnier/elementbyrecipe/command/infra/repository/CocktailInfraRepo.java
@@ -1,0 +1,4 @@
+package com.chainsmoker.marronnier.elementbyrecipe.command.infra.repository;
+
+public class CocktailInfraRepo {
+}
