+++ src/main/java/com/chainsmoker/marronnier/elementbyrecipe/command/infra/service/CocktailInfraService.java
@@ -1,0 +1,4 @@
+package com.chainsmoker.marronnier.elementbyrecipe.command.infra.service;
+
+public class CocktailInfraService {
+}
