+++ src/main/java/com/chainsmoker/marronnier/elementbyrecipe/query/application/controller/CocktailQController.java
@@ -1,0 +1,4 @@
+package com.chainsmoker.marronnier.elementbyrecipe.query.application.controller;
+
+public class CocktailQController {
+}
