+++ src/main/java/com/chainsmoker/marronnier/elementbyrecipe/query/application/dto/CocktailQDTO.java
@@ -1,0 +1,4 @@
+package com.chainsmoker.marronnier.elementbyrecipe.query.application.dto;
+
+public class CocktailQDTO {
+}
