+++ src/main/java/com/chainsmoker/marronnier/elementbyrecipe/query/application/service/CocktailQService.java
@@ -1,0 +1,4 @@
+package com.chainsmoker.marronnier.elementbyrecipe.query.application.service;
+
+public class CocktailQService {
+}
