+++ src/main/java/com/chainsmoker/marronnier/elementbyrecipe/query/domain/entity/ElementByRecipeQuery.java
@@ -1,0 +1,23 @@
+package com.chainsmoker.marronnier.elementbyrecipe.query.domain.entity;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class ElementByRecipeQuery {
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name="cocktail_id", nullable = false)
+    private Long cocktailId;
+
+    @Column(name="element_id",nullable = false)
+    private Long elementId;
+
+    @Column
+    private String capacity;
+}
+
