+++ src/main/java/com/chainsmoker/marronnier/elementbyrecipe/query/domain/repository/CocktailQRepo.java
@@ -1,0 +1,4 @@
+package com.chainsmoker.marronnier.elementbyrecipe.query.domain.repository;
+
+public class CocktailQRepo {
+}
