+++ src/main/java/com/chainsmoker/marronnier/elementbyrecipe/query/domain/service/CocktailQService.java
@@ -1,0 +1,4 @@
+package com.chainsmoker.marronnier.elementbyrecipe.query.domain.service;
+
+public class CocktailQService {
+}
