+++ src/main/java/com/chainsmoker/marronnier/elementbyrecipe/query/infra/repository/CocktailQInfraRepo.java
@@ -1,0 +1,4 @@
+package com.chainsmoker.marronnier.elementbyrecipe.query.infra.repository;
+
+public class CocktailQInfraRepo {
+}
