+++ src/main/java/com/chainsmoker/marronnier/elementbyrecipe/query/infra/service/CocktailQInfraService.java
@@ -1,0 +1,4 @@
+package com.chainsmoker.marronnier.elementbyrecipe.query.infra.service;
+
+public class CocktailQInfraService {
+}
